Move configuration out of API
