+++ src/api/IBMiContent.ts
@@ -7,7 +7,7 @@ import * as node_ssh from "node-ssh";
 import { GetMemberInfo } from './components/getMemberInfo';
 import { ObjectTypes } from '../filesystems/qsys/Objects';
 import { AttrOperands, CommandResult, IBMiError, IBMiMember, IBMiObject, IFSFile, QsysPath, SpecialAuthorities } from '../typings';
-import { ConnectionConfiguration } from './Configuration';
+import { ConnectionConfiguration } from '../Configuration';
 import { FilterType, parseFilter, singleGenericName } from './Filter';
 import { default as IBMi } from './IBMi';
 import { Tools } from './Tools';
