+++ src/views/LibraryListView.ts
@@ -1,5 +1,5 @@
 import vscode, { commands, l10n } from "vscode";
-import { ConnectionConfiguration, GlobalVSCodeConfiguration } from "../api/Configuration";
+import { ConnectionConfiguration, GlobalVSCodeConfiguration } from "../Configuration";
 import { instance } from "../instantiate";
 import { IBMiObject, WithLibrary } from "../typings";
 import { objectToToolTip } from "./tools";
