+++ src/views/Terminal.ts
@@ -2,7 +2,7 @@
 import path from 'path';
 import vscode, { commands } from 'vscode';
 import { instance } from '../instantiate';
-import { GlobalVSCodeConfiguration } from '../api/Configuration';
+import { GlobalVSCodeConfiguration } from '../Configuration';
 import IBMi from '../api/IBMi';
 import { Tools } from '../api/Tools';
