+++ src/views/objectBrowser.ts
@@ -2,7 +2,7 @@ import fs, { existsSync } from "fs";
 import os from "os";
 import path, { basename, dirname } from "path";
 import vscode from "vscode";
-import { ConnectionConfiguration, DefaultOpenMode, GlobalVSCodeConfiguration } from "../api/Configuration";
+import { ConnectionConfiguration, DefaultOpenMode, GlobalVSCodeConfiguration } from "../Configuration";
 import { parseFilter, singleGenericName } from "../api/Filter";
 import IBMi, { MemberParts } from "../api/IBMi";
 import { SortOptions, SortOrder } from "../api/IBMiContent";
