+++ src/webviews/variables/index.ts
@@ -1,5 +1,5 @@
 import vscode from "vscode";
-import { ConnectionConfiguration } from "../../api/Configuration";
+import { ConnectionConfiguration } from "../../Configuration";
 import { CustomUI } from "../CustomUI";
 import { instance } from "../../instantiate";
